refactor : shared ui button 변경
